refactor: extract useMindStateHold; document setLiveElapsed re-render trick
